+++ Thread/ThreadPri.h
@@ -1,5 +1,6 @@
 #ifndef __ThreadPri_H__
 #define __ThreadPri_H__
+#include <sys/types.h>
 
 class ThreadPri
 {
@@ -7,7 +8,7 @@ public:
 	virtual ~ThreadPri(){}
 	virtual bool start()=0;
 	virtual void stop()=0;
-	virtual unsigned GetThreadID()=0;
+	virtual size_t GetThreadID()=0;
 	virtual bool IsRuning()=0;
 };
